requirements.txt
langgraph>=0.1.0
langchain-core>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
groq>=0.4.0
python-dotenv>=1.0.0
